:wrench: Simplify scripts and development (#155)
* :wrench: Update scripts and configs

* :wrench: Add shebang to script

* :fire: Remove test and dev configs, pass inline

* :sparkles: Add local development dev-link set up

* :memo: Update generated docs with refactor

* :memo: Add Contributing guide

+++ scripts/dev-fsfp-back.sh
@@ -5,6 +5,11 @@
 # Exit in case of error
 set -e
 
+if [ ! -d ./full-stack-fastapi-postgresql ] ; then
+    echo "Run this script from outside the project, to integrate a sibling dev-fsfp project with changes and review modifications"
+    exit 1
+fi
+
 if [ $(uname -s) = "Linux" ]; then
     echo "Remove __pycache__ files"
     sudo find ./dev-fsfp/ -type d -name __pycache__ -exec rm -r {} \+
